+++ TintaExpressBackend/Models/materia_prima.cs
@@ -7,8 +7,8 @@ namespace TintaExpressBackend.Models
         [Key]
         public int id { get; set; }
         public string? nombre { get; set; }
-        public float inventario { get; set; }
+        public double inventario { get; set; }
         public string? unidad { get; set; }
-        public float minimo { get; set; }
+        public double minimo { get; set; }
     }
 }
